+++ packages/orq-rc/src/models/operations/getoneknowledge.ts
@@ -76,10 +76,6 @@ export type GetOneKnowledgeResponseBody = {
    */
   id: string;
   /**
-   * The id of the resource
-   */
-  createdById: string;
-  /**
    * The creation date of the knowledge base.
    */
   created: string;
@@ -103,10 +99,8 @@ export type GetOneKnowledgeResponseBody = {
    * The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
    */
   retrievalSettings?: GetOneKnowledgeRetrievalSettings | undefined;
-  /**
-   * The id of the resource
-   */
-  updateById: string;
+  createdById?: string | null | undefined;
+  updatedById?: string | null | undefined;
   /**
    * The last update date of the knowledge base.
    */
@@ -355,7 +349,6 @@ export const GetOneKnowledgeResponseBody$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   _id: z.string(),
-  created_by_id: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -364,28 +357,29 @@ export const GetOneKnowledgeResponseBody$inboundSchema: z.ZodType<
   retrieval_settings: z.lazy(() =>
     GetOneKnowledgeRetrievalSettings$inboundSchema
   ).optional(),
-  update_by_id: z.string(),
+  created_by_id: z.nullable(z.string()).optional(),
+  updated_by_id: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "_id": "id",
-    "created_by_id": "createdById",
     "retrieval_settings": "retrievalSettings",
-    "update_by_id": "updateById",
+    "created_by_id": "createdById",
+    "updated_by_id": "updatedById",
   });
 });
 
 /** @internal */
 export type GetOneKnowledgeResponseBody$Outbound = {
   _id: string;
-  created_by_id: string;
   created: string;
   description?: string | undefined;
   key: string;
   model: string;
   path?: string | undefined;
   retrieval_settings?: GetOneKnowledgeRetrievalSettings$Outbound | undefined;
-  update_by_id: string;
+  created_by_id?: string | null | undefined;
+  updated_by_id?: string | null | undefined;
   updated: string;
 };
 
@@ -396,7 +390,6 @@ export const GetOneKnowledgeResponseBody$outboundSchema: z.ZodType<
   GetOneKnowledgeResponseBody
 > = z.object({
   id: z.string(),
-  createdById: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -405,14 +398,15 @@ export const GetOneKnowledgeResponseBody$outboundSchema: z.ZodType<
   retrievalSettings: z.lazy(() =>
     GetOneKnowledgeRetrievalSettings$outboundSchema
   ).optional(),
-  updateById: z.string(),
+  createdById: z.nullable(z.string()).optional(),
+  updatedById: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     id: "_id",
-    createdById: "created_by_id",
     retrievalSettings: "retrieval_settings",
-    updateById: "update_by_id",
+    createdById: "created_by_id",
+    updatedById: "updated_by_id",
   });
 });