+++ packages/orq-rc/src/models/operations/listknowledgebases.ts
@@ -91,10 +91,6 @@ export type ListKnowledgeBasesData = {
    */
   id: string;
   /**
-   * The id of the resource
-   */
-  createdById: string;
-  /**
    * The creation date of the knowledge base.
    */
   created: string;
@@ -118,10 +114,8 @@ export type ListKnowledgeBasesData = {
    * The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
    */
   retrievalSettings?: ListKnowledgeBasesRetrievalSettings | undefined;
-  /**
-   * The id of the resource
-   */
-  updateById: string;
+  createdById?: string | null | undefined;
+  updatedById?: string | null | undefined;
   /**
    * The last update date of the knowledge base.
    */
@@ -411,7 +405,6 @@ export const ListKnowledgeBasesData$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   _id: z.string(),
-  created_by_id: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -420,28 +413,29 @@ export const ListKnowledgeBasesData$inboundSchema: z.ZodType<
   retrieval_settings: z.lazy(() =>
     ListKnowledgeBasesRetrievalSettings$inboundSchema
   ).optional(),
-  update_by_id: z.string(),
+  created_by_id: z.nullable(z.string()).optional(),
+  updated_by_id: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "_id": "id",
-    "created_by_id": "createdById",
     "retrieval_settings": "retrievalSettings",
-    "update_by_id": "updateById",
+    "created_by_id": "createdById",
+    "updated_by_id": "updatedById",
   });
 });
 
 /** @internal */
 export type ListKnowledgeBasesData$Outbound = {
   _id: string;
-  created_by_id: string;
   created: string;
   description?: string | undefined;
   key: string;
   model: string;
   path?: string | undefined;
   retrieval_settings?: ListKnowledgeBasesRetrievalSettings$Outbound | undefined;
-  update_by_id: string;
+  created_by_id?: string | null | undefined;
+  updated_by_id?: string | null | undefined;
   updated: string;
 };
 
@@ -452,7 +446,6 @@ export const ListKnowledgeBasesData$outboundSchema: z.ZodType<
   ListKnowledgeBasesData
 > = z.object({
   id: z.string(),
-  createdById: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -461,14 +454,15 @@ export const ListKnowledgeBasesData$outboundSchema: z.ZodType<
   retrievalSettings: z.lazy(() =>
     ListKnowledgeBasesRetrievalSettings$outboundSchema
   ).optional(),
-  updateById: z.string(),
+  createdById: z.nullable(z.string()).optional(),
+  updatedById: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     id: "_id",
-    createdById: "created_by_id",
     retrievalSettings: "retrieval_settings",
-    updateById: "update_by_id",
+    createdById: "created_by_id",
+    updatedById: "updated_by_id",
   });
 });