+++ packages/orq-rc/src/models/operations/updateknowledge.ts
@@ -147,10 +147,6 @@ export type UpdateKnowledgeResponseBody = {
    */
   id: string;
   /**
-   * The id of the resource
-   */
-  createdById: string;
-  /**
    * The creation date of the knowledge base.
    */
   created: string;
@@ -174,10 +170,8 @@ export type UpdateKnowledgeResponseBody = {
    * The retrieval settings for the knowledge base. If not provider, Hybrid Search will be used as a default query strategy.
    */
   retrievalSettings?: UpdateKnowledgeKnowledgeRetrievalSettings | undefined;
-  /**
-   * The id of the resource
-   */
-  updateById: string;
+  createdById?: string | null | undefined;
+  updatedById?: string | null | undefined;
   /**
    * The last update date of the knowledge base.
    */
@@ -699,7 +693,6 @@ export const UpdateKnowledgeResponseBody$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   _id: z.string(),
-  created_by_id: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -708,21 +701,21 @@ export const UpdateKnowledgeResponseBody$inboundSchema: z.ZodType<
   retrieval_settings: z.lazy(() =>
     UpdateKnowledgeKnowledgeRetrievalSettings$inboundSchema
   ).optional(),
-  update_by_id: z.string(),
+  created_by_id: z.nullable(z.string()).optional(),
+  updated_by_id: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     "_id": "id",
-    "created_by_id": "createdById",
     "retrieval_settings": "retrievalSettings",
-    "update_by_id": "updateById",
+    "created_by_id": "createdById",
+    "updated_by_id": "updatedById",
   });
 });
 
 /** @internal */
 export type UpdateKnowledgeResponseBody$Outbound = {
   _id: string;
-  created_by_id: string;
   created: string;
   description?: string | undefined;
   key: string;
@@ -731,7 +724,8 @@ export type UpdateKnowledgeResponseBody$Outbound = {
   retrieval_settings?:
     | UpdateKnowledgeKnowledgeRetrievalSettings$Outbound
     | undefined;
-  update_by_id: string;
+  created_by_id?: string | null | undefined;
+  updated_by_id?: string | null | undefined;
   updated: string;
 };
 
@@ -742,7 +736,6 @@ export const UpdateKnowledgeResponseBody$outboundSchema: z.ZodType<
   UpdateKnowledgeResponseBody
 > = z.object({
   id: z.string(),
-  createdById: z.string(),
   created: z.string(),
   description: z.string().optional(),
   key: z.string(),
@@ -751,14 +744,15 @@ export const UpdateKnowledgeResponseBody$outboundSchema: z.ZodType<
   retrievalSettings: z.lazy(() =>
     UpdateKnowledgeKnowledgeRetrievalSettings$outboundSchema
   ).optional(),
-  updateById: z.string(),
+  createdById: z.nullable(z.string()).optional(),
+  updatedById: z.nullable(z.string()).optional(),
   updated: z.string(),
 }).transform((v) => {
   return remap$(v, {
     id: "_id",
-    createdById: "created_by_id",
     retrievalSettings: "retrieval_settings",
-    updateById: "update_by_id",
+    createdById: "created_by_id",
+    updatedById: "updated_by_id",
   });
 });